refactor: refactor notification service for improved performance
- Add `google.golang.org/grpc/metadata` to the imports in `notification_grpc_server.go`
- Change `Limit` and `Offset` types to `int64` in `notification.go` file
- Set default `Limit` value to `100` in `mongodb.go` file

+++ app/domain/notification/biz/notification.go
@@ -45,8 +45,8 @@ func (i *impl) ListNotificationsByUser(
 	defer span.End()
 
 	return i.notifications.List(ctx, repo.ListCondition{
-		Limit:  options.Size,
-		Offset: (options.Page - 1) * options.Size,
+		Limit:  int64(options.Size),
+		Offset: int64((options.Page - 1) * options.Size),
 		UserID: userID,
 	})
 }
