+++ entity/domain/notification/repo/notification.go
@@ -9,8 +9,8 @@ import (
 
 // ListCondition defines the conditions for listing notifications.
 type ListCondition struct {
-	Limit  int
-	Offset int
+	Limit  int64
+	Offset int64
 
 	UserID string
 }
